src/network-monitoring/requirements.txt
# Core dependencies
prometheus-client>=0.19.0
psutil>=5.9.0
requests>=2.31.0
pyyaml>=6.0
jinja2>=3.1.0
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
